+++ projects/example/src/app/app.module.ts
@@ -9,7 +9,7 @@ import { AppRatingComponent } from './rating.component';
 @NgModule({
   declarations: [
     AppComponent,
-    AppRatingComponent
+    AppRatingComponent,
   ],
   imports: [
     BrowserModule,
@@ -26,13 +26,13 @@ import { AppRatingComponent } from './rating.component';
           itemsPerPageLabel: {
             singular: 'film per pagina',
             plural: 'films per pagina',
-          }
-        }
-      })
+          },
+        },
+      }),
   ],
   providers: [],
   bootstrap: [AppComponent],
-  entryComponents: [AppRatingComponent]
+  entryComponents: [AppRatingComponent],
 })
 export class AppModule {
 }