+++ projects/smart-table/src/lib/components/table-datepicker-filter/table-datepicker-filter.component.ts
@@ -5,7 +5,7 @@ import { AbstractFilterComponent } from '../filter/abstract-filter-component';
 @Component({
   selector: 'aui-table-datepicker-filter',
   templateUrl: 'table-datepicker-filter.component.html',
-  styleUrls: ['../filter/filter.component.scss']
+  styleUrls: ['../filter/filter.component.scss'],
 })
 export class TableDatepickerFilterComponent extends AbstractFilterComponent implements OnInit {
   ngOnInit() {