+++ projects/smart-table/src/lib/smart-table.types.ts
@@ -10,7 +10,8 @@ export interface IModuleConfig {
   options?: {
     useLowerCaseQueryParams?: boolean,
     noConfigApiCall?: boolean,
-    noExport?: boolean
+    noExport?: boolean,
+    exportWithFilters?: boolean,
   };
 }
 
@@ -86,7 +87,7 @@ export enum SmartTableColumnType {
   Number = 'number',
   Text = 'text',
   Date = 'date',
-  DateTime = 'dateTime'
+  DateTime = 'dateTime',
 }
 
 export interface SmartTableColumnCustomType {
@@ -99,19 +100,19 @@ export enum SmartTableFilterType {
   Select = 'select',
   Input = 'input',
   Datepicker = 'datepicker',
-  SearchFilter = 'search-filter'
+  SearchFilter = 'search-filter',
 }
 
 export enum SmartTableFilterDisplay {
   Generic = 'generic',
   Visible = 'visible',
-  Optional = 'optional'
+  Optional = 'optional',
 }
 
 export enum SmartTableFilterOperator {
   Equal = '=',
   ILike = 'ILIKE',
-  In = 'in'
+  In = 'in',
 }
 
 export interface SmartTableFilterConfig {